(org) Rename canonical to canonicity
- move Canonicity to canonicity mod
