--- src/canonical/mod.rs
@@ -1,2 +1,0 @@
-pub mod chain_discovery;
-pub mod store;
