+++ src/ipc.rs
@@ -7,7 +7,7 @@ use tracing::{debug, error, info, instrument, trace};
 
 use crate::{
     block::{store::BlockStore, Block, BlockHash},
-    canonical::store::CanonicityStore,
+    canonicity::store::CanonicityStore,
     ledger::{public_key::PublicKey, store::LedgerStore, Ledger},
     server::{IndexerConfiguration, IpcChannelUpdate},
     state::summary::{SummaryShort, SummaryVerbose},
